ATR-588: renamed namespace, folder + minor refactoring
